Fix: notiGateway -> websocket-listening으로 분리
