+++ src/notifications/notifications.module.ts
@@ -1,12 +1,9 @@
-import { Module, forwardRef } from '@nestjs/common';
-import { NotificationsGateway } from './notifications.gateway';
-import { UserRelationModule } from 'src/user-relation/user-relation.module';
-import { ChannelsModule } from 'src/channels/channels.module';
-import { CommonsModule } from 'src/commons/commons.module';
+import { Module } from '@nestjs/common';
+import { NotificationsEmitGateway } from './notifications-emit.gateway';
 
 @Module({
-  imports: [forwardRef(() => UserRelationModule), forwardRef(() => ChannelsModule), CommonsModule],
-  providers: [NotificationsGateway],
-  exports: [NotificationsGateway],
+  imports: [],
+  providers: [NotificationsEmitGateway],
+  exports: [NotificationsEmitGateway],
 })
 export class NotificationsModule {}
